Move testing compatibility stubs from _micropip into separate files
